install k6 from smoketest gha and run with installed executable

+++ run-k6-smoketests.sh
@@ -14,20 +14,29 @@ fi
 
 echo "starting k6"
 echo "running k6 frontend tests"
-docker run --rm -v $(pwd)/dpc-load-testing:/src --env-file $(pwd)/ops/config/decrypted/local.env \
-  --network host \
+k6 run \
   -e PORTAL_HOST=${PORTAL_HOST} \
   -e WEB_ADMIN_HOST=${WEB_ADMIN_HOST} \
   -e WEB_HOST=${WEB_HOST} \
-  -i grafana/k6 run /src/smoketest-frontend-ci-app.js
+  ./dpc-load-testing/smoketest-frontend-ci-app.js
 
-echo "running (authenticated) backend tests in docker"
-docker run --rm -v $(pwd)/dpc-load-testing:/src --env-file $(pwd)/ops/config/decrypted/local.env \
-  --network host \
-  -e PORTAL_HOST=${PORTAL_HOST} \
-  -e WEB_ADMIN_HOST=${WEB_ADMIN_HOST} \
-  -e WEB_HOST=${WEB_HOST} \
-  -e API_ADMIN_URL=${API_ADMIN_URL} \
-  -e ENVIRONMENT=${K6_ENVIRONMENT} -i grafana/k6 run /src/smoketest-backend-ci-app.js
+if [[ "$*" == *"--k6-env=local"* ]]; then
+  echo "running (authenticated) backend tests in docker locally"
+  docker run --rm -v $(pwd)/dpc-load-testing:/src --env-file $(pwd)/ops/config/decrypted/local.env \
+    --network host \
+    -e PORTAL_HOST=${PORTAL_HOST} \
+    -e WEB_ADMIN_HOST=${WEB_ADMIN_HOST} \
+    -e WEB_HOST=${WEB_HOST} \
+    -e API_ADMIN_URL=${API_ADMIN_URL} \
+    -e ENVIRONMENT=${K6_ENVIRONMENT} -i grafana/k6 run /src/smoketest-backend-ci-app.js
+else
+  echo "running (authenticated) backend tests against ${HOST_URL}"
+  k6 run \
+    -e PORTAL_HOST=${PORTAL_HOST} \
+    -e WEB_ADMIN_HOST=${WEB_ADMIN_HOST} \
+    -e WEB_HOST=${WEB_HOST} \
+    -e API_ADMIN_URL=${API_ADMIN_URL} \
+    -e ENVIRONMENT=${K6_ENVIRONMENT} ./dpc-load-testing/smoketest-backend-ci-app.js
+fi
 
 echo "Done!"
